+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 ### Breaking changes
 
 - `sandbox#touchp()` and `sandbox#mkdirp()` now also accept an array of paths and create and return all specified files/directories.
+
+- files now have content